fix(native-compiler): Phase 0.6 review fixes — escaping, export const, fraction parsing (#1871)
Fixes 3 blockers and 3 should-fix items from adversarial review:

- B-1: Fix string escaping order (backslashes before quotes)
- B-2: Handle `export const X = createContext()` via ExportNamedDeclaration unwrap
- B-3: Accept hash algorithm divergence (SipHash vs Wyhash) as known difference
- SF-1: Remove const-only restriction on createContext (match TS behavior)
- SF-2: Restrict CSS fraction parsing to integers (match TS regex)
- SF-3: Add test for export const createContext pattern

184 tests passing, clippy clean.

+++ native/vertz-compiler/src/css_token_tables.rs
@@ -430,13 +430,13 @@ fn resolve_size(value: &str, property: &str) -> Option<String> {
     if let Some(v) = size_keywords(value) {
         return Some(v.to_string());
     }
-    // Fraction: N/M → percentage
+    // Fraction: N/M → percentage (integers only, matching TS regex /^(\d+)\/(\d+)$/)
     if let Some(slash_idx) = value.find('/') {
         let num_str = &value[..slash_idx];
         let den_str = &value[slash_idx + 1..];
-        if let (Ok(num), Ok(den)) = (num_str.parse::<f64>(), den_str.parse::<f64>()) {
-            if den != 0.0 {
-                let pct = (num / den) * 100.0;
+        if let (Ok(num), Ok(den)) = (num_str.parse::<u64>(), den_str.parse::<u64>()) {
+            if den != 0 {
+                let pct = (num as f64 / den as f64) * 100.0;
                 if pct % 1.0 == 0.0 {
                     return Some(format!("{}%", pct as i64));
                 }
